rtai_math: get rid of use of userspace headers in kernel modules

+++ base/math/s_scalbn.c
@@ -17,7 +17,6 @@
  * exponentiation or a multiplication.
  */
 
-#include "math.h"
 #include "mathP.h"
 
 static const double
@@ -26,7 +25,7 @@ twom54  =  5.55111512312578270212e-17, /* 0x3C900000, 0x00000000 */
 huge   = 1.0e+300,
 tiny   = 1.0e-300;
 
-	double scalbn (double x, int n)
+	double scalbn (double x, long int n)
 {
 	int32_t k,hx,lx;
 	EXTRACT_WORDS(hx,lx,x);
